Removed my spEntities and added the SP List entities to the existing SPEntities file.
