--- src/common/spEntities/ISPList.ts
@@ -1,8 +1,0 @@
-export interface ISPList {
-    Id: string;
-    Title: string;
-    BaseTemplate: string;
-}
-export interface ISPLists {
-    value: ISPList[];
-}
