--- src/common/spEntities/index.ts
@@ -1,1 +1,0 @@
-export * from './ISPList';
